hosting partials
